+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Allow bindout in embedded and immediate sql
 * Added an activity burst in oltp workload
 * Save CPU using faster code to convert integer/double to string
+* Add normalrandom function returning normal distributed random numbers
 
 ## 3.0.5